+++ compiler-tests/src/test/java/dev/zacsweers/metro/compiler/BoxTestGenerated.java
@@ -1071,6 +1071,12 @@ public class BoxTestGenerated extends AbstractBoxTest {
     }
 
     @Test
+    @TestMetadata("ContributedAccessorsCanBeLazy.kt")
+    public void testContributedAccessorsCanBeLazy() {
+      run("ContributedAccessorsCanBeLazy.kt");
+    }
+
+    @Test
     @TestMetadata("ContributesCanBeInject.kt")
     public void testContributesCanBeInject() {
       run("ContributesCanBeInject.kt");